refactoring with webpack+typescript+eslint
