+++ carlo.js
@@ -1,12 +1,19 @@
 const carlo = require('carlo');
 const AWS = require('aws-sdk');
+const fs = require('fs');
 
 (async () => {
   const app = await carlo.launch();
   let closing = false;
   app.on('exit', () => {
-    closing = true
-    process.exit()
+    closing = true;
+    process.exit();
+  });
+  app.serveHandler(request => {
+    if (request.url() === 'https://localhost/aws-step-functions.js')
+      request.fulfill({body: fs.readFileSync('./dist/aws-step-functions.js')});
+    else
+      request.continue();
   });
   app.serveFolder(__dirname);
 
@@ -14,41 +21,41 @@ const AWS = require('aws-sdk');
 
   await app.exposeFunction('__updateAWSconfig', params => {
     AWS.config.update(params);
-    return Promise.resolve()
+    return Promise.resolve();
   });
 
   await app.exposeFunction('__listStateMachines', async _ => {
     try {
       const stepfunctions = new AWS.StepFunctions({apiVersion: '2016-11-23'});
-      return await stepfunctions.listStateMachines({}).promise()
+      return await stepfunctions.listStateMachines({}).promise();
     } catch (err) {
-      console.log(err)
-      throw err
+      console.log(err);
+      throw err;
     }
   });
 
   await app.exposeFunction('__describeStateMachine', async arn => {
     try {
       const stepfunctions = new AWS.StepFunctions({apiVersion: '2016-11-23'});
-      return await stepfunctions.describeStateMachine({ stateMachineArn: arn }).promise()
+      return await stepfunctions.describeStateMachine({ stateMachineArn: arn }).promise();
     } catch (err) {
-      console.log(err)
-      throw err
+      console.log(err);
+      throw err;
     }
-  })
+  });
 
   await app.exposeFunction('__deployStateMachine', async params => {
     try {
       const stepfunctions = new AWS.StepFunctions({apiVersion: '2016-11-23'});
       if (params.stateMachineArn) {
-        return await stepfunctions.updateStateMachine(params).promise()
+        return await stepfunctions.updateStateMachine(params).promise();
       } else {
-        return await stepfunctions.createStateMachine(params).promise()
+        return await stepfunctions.createStateMachine(params).promise();
       }
     } catch (err) {
-      console.log(err)
-      throw err
+      console.log(err);
+      throw err;
     }
-  })
+  });
 
-})()
+})();
