+++ src/@types/awssf.d.ts
@@ -1,0 +1,41 @@
+declare module AWS {
+  const config
+  const STS
+  const Lambda
+}
+
+declare module jsyaml {
+  function load(data: string)
+  function dump(obj)
+}
+declare function jsep(condition)
+
+declare let AWSconfig
+declare let StartPoint
+declare let EndPoint
+declare let TaskState
+declare let PassState
+declare let ChoiceState
+declare let WaitState
+declare let SucceedState
+declare let FailState
+declare let ParallelState
+
+declare let StartAtEdge
+declare let NextEdge
+declare let RetryEdge
+declare let ChoiceEdge
+declare let CatchEdge
+declare let DefaultEdge
+
+interface StepFunctionsDefinitions {
+  Comment?: string
+  StartAt: string
+  States: any
+  TimeoutSeconds?: number
+  Version?: string
+}
+
+interface mxCell {
+  awssf: any
+}
