+++ src/@types/carlo.d.ts
@@ -1,0 +1,4 @@
+declare function __listStateMachines(): Promise<any>
+declare function __describeStateMachine(params): Promise<any>
+declare function __updateAWSconfig(params): Promise<any>
+declare function __deployStateMachine(params): Promise<any>
