+++ src/index.ts
@@ -1,0 +1,10 @@
+mxscript("https://cdn.rawgit.com/soney/jsep/b8baab7b/build/jsep.min.js");
+mxscript("https://cdn.rawgit.com/nodeca/js-yaml/9c1894e2/dist/js-yaml.min.js");
+mxscript("https://sdk.amazonaws.com/js/aws-sdk-2.510.0.min.js");
+// mxscript("https://rawgit.com/soney/jsep/master/build/jsep.min.js");
+// mxscript("https://f248fda6.ngrok.io/aws-sdk-2.510.0.js");
+// mxscript("https://localhost:8000/js/jsep.min.js");
+
+import main from "./main";
+
+Draw.loadPlugin(main);
