+++ src/patch.ts
@@ -1,0 +1,9 @@
+export function refreshScratchpadData (ui) {
+  var forceReloadScratchPad = function (...v) {
+    console.log(">>reload scratchpad", ...v);
+    ui.toggleScratchpad();
+    ui.toggleScratchpad();
+  };
+  
+  ui.addListener('clientLoaded', forceReloadScratchPad);
+}
